Refactor core crypto package
