+++ pkg/auth/handler/sso/auth_url.go
@@ -14,9 +14,9 @@ import (
 	"github.com/skygeario/skygear-server/pkg/core/auth/authz"
 	"github.com/skygeario/skygear-server/pkg/core/auth/authz/policy"
 	"github.com/skygeario/skygear-server/pkg/core/config"
+	"github.com/skygeario/skygear-server/pkg/core/crypto"
 	"github.com/skygeario/skygear-server/pkg/core/db"
 	"github.com/skygeario/skygear-server/pkg/core/handler"
-	"github.com/skygeario/skygear-server/pkg/core/hash"
 	coreHttp "github.com/skygeario/skygear-server/pkg/core/http"
 	"github.com/skygeario/skygear-server/pkg/core/inject"
 	"github.com/skygeario/skygear-server/pkg/core/server"
@@ -225,7 +225,7 @@ func (h *AuthURLHandler) Handle(w http.ResponseWriter, r *http.Request) (result 
 				UXMode:      payload.UXMode,
 				Action:      h.Action,
 			},
-			Nonce: hash.SHA256String(nonce),
+			Nonce: crypto.SHA256String(nonce),
 		},
 	}
 	if h.AuthContext.AuthInfo() != nil {
