+++ pkg/core/crypto/sha256.go
@@ -1,4 +1,4 @@
-package hash
+package crypto
 
 import (
 	"crypto/sha256"
