+++ pkg/core/crypto/sha256_test.go
@@ -1,4 +1,4 @@
-package hash
+package crypto
 
 import (
 	"testing"
